(#7511) Add --region flag to support EC2 zones
This change allows Cloud Pack to operate in any of the regions Amazon
EC2 operates in.  These regions are currently eu-west-1, us-east-1,
ap-northeast-1, us-west-1, ap-southeast-1

The option defaults to 'us-east-1' if the end user does not specify the
--region option.

This option is only valid for the create and terminate actions of Cloud
Pack.

The rspec tests have been updated to validate --region on the create and
terminate actions.