(#8185) Add a central place to set option defaults
For #8185 I'm adding a list action.  This action does not require any
options, but the Fog AWS call needs to have at least the platform option
set to 'AWS'  In addition, the region option is optional for Fog but
should default to 'us-east-1' to prevent Puppet's behavior from changing
if Fog's default region changes.

This patch adds a new method, #merge_default_options which sets the
default platform to AWS and default region to us-east-1.  Any actions
that send a message to #create_connection should run their options hash
through the merge_default_options and re-assign.

This is particularly important for the create, terminate and upcoming
list actions.