(#8185) Add list action
This change adds a new action to cloud pack.  This increases the number
of actions that are EC2 specific.  Previously, it was just create and
terminate.

The Faces API is responsible for rendering the output of this action.  A
simple array of DNS names, one for each running instance in the EC2
region is handed back to Faces.

Sample output:

    $ puppet cloudnode list --region us-west-1 --render-as yaml
    --- []
    $ puppet cloudnode list
    [null]
    $ puppet cloudnode list --render-as yaml
    puppet cloudnode list --render-as yaml
    ---
      - ec2-50-19-71-40.compute-1.amazonaws.com

Spec tests have also been added validating the input options and output
of the action.