(#8343) Create action prints dns name asap
This small change gives the end user useful information about the
instance ID and the public DNS name as soon as they are available inside
the create action.  Without this change, the end user may not have the
information they need at the time they need it.

This also allows the end user to "do stuff" while waiting for the
fingerprints.

Here is the new output:

    $ puppet cloudnode create --image ami-2342a94a \
      --keypair jeff_initial --type t1.micro
    Connecting to AWS us-east-1 Done
    Instance Type: t1.micro
    Creating new instance ... Done
    Instance identifier: i-a80be6c9
    Creating tags for instance ... Done
    Starting up ............. Done
    Instance public dns name: ec2-50-17-178-240.compute-1.amazonaws.com
    Waiting for host fingerprints .....................................