(#8350) Change all output to use logdest
This change replaces all instances of print and puts with
Puppet.{info,warning,err}().  This change is motivated by two main
reasons:  First, the standard behavior of the Faces API is to take the
return value of an action and render the object to STDOUT for use in
pipelines with other actions.

Second, the Puppet end user already has an expectation that
informational messages in Puppet are configurable using the log
destination feature.

Note, this change makes all of the cloudpack output and feedback silent unless
the end user passes the --verbose flag.

In addition, I disabled the fingerprint spec test checking if the finger
print is printed to standard output.  The test was passing before this
change and I've never seen the fingerprint printed to STDOUT so I don't
think it was working anyway.  In addition, #8348 proposes fingerprinting
will be it's own action which should simply return a string to the API
for rendering to the user.  The spec test should be unmarked as pending
in that ticket.

Reviewed-by: James Turnbull