(#8350) Change progress to a bar to avoid logspam
In an effort to avoid log spam but also provide the end user interactive
feedback, this change updates the shift from print and puts to only use
print when outputting a progress bar.

Hopefully, # will be interpreted as a comment in YAML renderings, and
the user may filter all lines starting with # before creating a JSON
pipeline.

Reviewed-by: James Turnbull